Running pychecker over ZSI, and I noticed some problems in wstools.
I fixed that fact that Notation and Entity were not be found. I changed them
to use the qualified name like the rest of the symbols from xml.dom.minidom.

I also discovered that a RecursionError was being thrown, but RecursionError
didn't exist. I created simple sub-class of Exception to rectify this.


git-svn-id: https://pywebsvcs.svn.sourceforge.net/svnroot/pywebsvcs/trunk@248 7150bf37-e60d-0410-b93f-83e91ef0e581