remove field lenofwhat from Struct base class.
this field was only used in the parse() method to sanity test the response from
a server.  the code compared the length of the typecode list with the length
of the elements received in a response from the server.  if the two equal,
parse.  if not equal, barf.

there is an interesting circular dependency trick in existence with typecode
rewriting, which unfortunately does not work with the TCcompound code in it's
current state.  the crux of the issue is that when TCcompound loads the
complex types module provided by a developer, the typecode does not possess
the circular reference, thereby throwing off the lenofwhat.

*but*, by the time parse is called, the complex type has the correct length.

so, this patch changes the two references to self.lenofwhat to use
len(self.ofwhat) instead.


git-svn-id: https://pywebsvcs.svn.sourceforge.net/svnroot/pywebsvcs/trunk@621 7150bf37-e60d-0410-b93f-83e91ef0e581