Fixed bug in code used with use_wsdl=True argument is passed to ServiceProxy constructor.
The bug that existed was that the code was ignoring minOccurs and maxOccurs
fields on any types that were not primitives.  If you had:
  A := B (maxOccurs="unbounded")
  B := C, D (C and D both are type xsd:string, and both "unbounded")

The code was only setting the repeatable flag for the typecode generated for
C and D.  The typecode generated for A did not contain the repeatable flag,
nor did it include the optional flag (if minOccurs="0").

I changed the code to honor minOccurs and maxOccurs at intermediate levels of
the typecode generation.


git-svn-id: https://pywebsvcs.svn.sourceforge.net/svnroot/pywebsvcs/trunk@673 7150bf37-e60d-0410-b93f-83e91ef0e581