M    ZSI/dispatch.py - fixed a bug, expecting a single instance  not a sequence w/non rpc dispatching M    ZSI/TC.py - make soap:fault wildcard lax processing (just want to dump whatever in it, disregard schema definition) M    ZSI/fault.py - change a logger.error message to logger.debug
This was a little tricky.  Two problems, first a little bug passing the parsed request as *arg (it was an empty string), causing a soap fault to be generated which uses "AnyElement" which causes the problem because the rules for processing were tightened up.


So when I fixed the soap fault problem, it returns:

$ python client.py
Traceback (most recent call last):
  File "client.py", line 10, in ?
    response = port.getSequenceAnnot(req)
  File "/Users/boverhof/Desktop/Test/Andreas/Tutorial/SequenceAnnotService_services.py", line 39, in getSequenceAnnot
    response = self.binding.Receive(tutorialResponse.typecode)
  File "/Users/boverhof/Desktop/Test/Andreas/Tutorial/client.py", line 502, in Receive
   
  File "/Users/boverhof/Desktop/Test/Andreas/Tutorial/client.py", line 430, in Receive
   
ZSI.FaultException: Unparseable message
<ZSI:ParseFaultDetail>
<ZSI:string>id() takes exactly 1 argument (0 given)</ZSI:string>
<ZSI:trace></ZSI:trace>
</ZSI:ParseFaultDetail>


Now the error at least describes points at the problem, which is easily fixed.

$ python client.py
GAG


git-svn-id: https://pywebsvcs.svn.sourceforge.net/svnroot/pywebsvcs/trunk@1293 7150bf37-e60d-0410-b93f-83e91ef0e581